add dry run flag and exclude flag

+++ src/index.ts
@@ -1,7 +1,7 @@
 import { Command, flags } from "@oclif/command";
 import { promisify } from "util";
 import { exec } from "child_process";
-import { appendFile, createReadStream, writeFile } from "fs-extra";
+import { appendFile, createReadStream, readFile, writeFile } from "fs-extra";
 import * as readline from 'readline'
 const faker = require("faker");
 const path = require("path");
@@ -33,7 +33,7 @@ class PgAnonymizer extends Command {
     help: flags.help({ char: "h" }),
     list: flags.string({
       char: "l",
-      description: "list of columns to anonymize",
+      description: "list of columns to anonymize ... or path to list file (no commas)",
       default:
         "email,name,description,address,city,country,phone,comment,birthdate",
     }),
@@ -53,6 +53,14 @@ class PgAnonymizer extends Command {
     inputFile: flags.string({
       char: "i",
       description: "input file (uses an input file instead of calling pg_dump)",
+    }),
+    dryRun: flags.boolean({
+      char: "d",
+      description: "dry run does not create actual files, just figures out what columns to process.",
+    }),
+    exclude: flags.string({
+      char: "x",
+      description: "tables to exclude data for (table structure is created, but data is not copied).",
     }),
     pgDumpOutputMemory: flags.string({
       char: "m",
@@ -99,7 +107,23 @@ class PgAnonymizer extends Command {
       // );
     }
 
-    const list = flags.list.split(",").map((l) => {
+    let excludeSourceString = flags.exclude;
+    let excludeTables: Set<string> = new Set();
+
+    if (excludeSourceString) {
+      if (!excludeSourceString.includes(',')) {
+        excludeSourceString = (await readFile(excludeSourceString)).toString().split('\n').join(',');
+      }
+      excludeTables = new Set(excludeSourceString.split(','));
+    }
+
+    let listSourceString = flags.list;
+
+    if (!listSourceString.includes(',')) {
+      listSourceString = (await readFile(listSourceString)).toString().split('\n').join(',');
+    }
+
+    const list = listSourceString.split(",").map((l) => {
       return {
         col: l.replace(/:(?:.*)$/, "").toLowerCase(),
         replacement: l.includes(":") ? l.replace(/^(?:.*):/, "") : null,
@@ -107,12 +131,20 @@ class PgAnonymizer extends Command {
     });
 
     let table = null;
-    let indices: Number[] = [];
+    let indices: Set<number> = new Set();
     let cols: string[] = [];
 
     console.log("Starting anonymization.");
-    console.log("Output file: " + flags.output);
-    await writeFile(flags.output, "");
+
+    if (flags.dryRun) {
+      console.log('DRY RUN MODE. Not creating file.');
+    } else {
+      console.log("Output file: " + flags.output);
+      await writeFile(flags.output, "");
+    }
+
+    console.log('   ');
+    console.log('   ');
 
     const fileStream = createReadStream(flags.inputFile);
 
@@ -121,10 +153,14 @@ class PgAnonymizer extends Command {
       crlfDelay: Infinity
     }) as any as Iterable<String>;
 
+    const allColsToAnonymize: Set<string> = new Set();
+    const allColsToNotAnonymize: Set<string> = new Set();
+
     for await (let line of inputLineResults) {
       // for (let line of result.split("\n")) {
       if (line.match(/^COPY .* FROM stdin;$/)) {
         table = line.replace(/^COPY (.*?) .*$/, "$1");
+        console.log('  ');
         console.log("Anonymizing table " + table);
 
         cols = line
@@ -134,22 +170,40 @@ class PgAnonymizer extends Command {
           .map((e) => e.replace(/"/g, ""))
           .map((e) => e.toLowerCase());
 
-        indices = cols.reduce((acc: Number[], value, key) => {
+        indices = new Set(cols.reduce((acc: number[], value, key) => {
           if (list.find((l) => l.col === value)) acc.push(key);
           return acc;
-        }, []);
+        }, []));
 
-        if (indices.length) 
-          console.log(
-            "Columns to anonymize: " +
-            cols.filter((v, k) => indices.includes(k)).join(", ")
-          );
-        else console.log("No columns to anonymize");
+        const colsToAnonymize: string[] = [];
+        const colsToNotAnonymize: string[] = [];
+
+        if (indices.size) {
+          cols.forEach((v, i) => {
+            if (indices.has(i)) {
+              colsToAnonymize.push(v);
+              allColsToAnonymize.add(v);
+            } else {
+              colsToNotAnonymize.push(v);
+              allColsToNotAnonymize.add(v);
+            }
+          })
+          console.log("Columns to anonymize: " + colsToAnonymize.join(", "));
+          console.log('NOT anonymize: ' + colsToNotAnonymize.join(", "));
+        } else {
+          console.log("No columns to anonymize");
+          console.log('NOT anonymize: ' + cols.join(', '));
+          console.log('  ');
+        }
       } else if (table && line.trim()) {
+        if (flags.dryRun || excludeTables.has(table)) {
+          continue;
+        }
+        
         line = line
           .split("\t")
           .map((v, k) => {
-            if (indices.includes(k)) {
+            if (indices.has(k)) {
               const replacement = list.find(
                 (l) => l.col === cols[k]
               )?.replacement;
@@ -189,15 +243,33 @@ class PgAnonymizer extends Command {
           .join("\t");
       } else {
         table = null;
-        indices = [];
+        indices = new Set();
         cols = [];
       }
+
+      if (flags.dryRun) {
+        continue;
+      }
+
       try {
         await appendFile(flags.output, line + "\n");
       } catch (e) {
         dieAndLog("Failed to write file", e);
       }
     }
+
+    console.log('  ');
+    console.log('  ');
+    console.log('Done.');
+    console.log('  ');
+    console.log('  ');
+
+    console.log('Summary');
+    console.log('  ');
+    console.log("Columns to anonymize: " + Array.from(allColsToAnonymize).join(", "));
+    console.log('  ');
+    console.log('NOT anonymize: ' + Array.from(allColsToNotAnonymize).join(", "));
+    console.log('  ');
   }
 }
